v2 for Assignment5
so... after talking to some people it appears this was the "correct" way to do it. Thankfully it didn't take as long as the other way... but i feel like i wasted so much time XDD. All of your code is used here, we just have to finalize the other things yall are working on and set the default page to login.